Revert Snowflake ID migration
Snowflake IDs exceed JavaScript's max safe number (2^53). This causes precision loss during JSON serialization, resulting in corrupted IDs in API payloads.

+++ cmd/internal/domain/entity/note.go
@@ -17,10 +17,10 @@ const (
 )
 
 type Note struct {
-	ID          int64          `gorm:"primaryKey"`
+	ID          int            `gorm:"primaryKey"`
 	Name        string         `gorm:"not null"`
 	Content     string         `gorm:"not null"`
-	CreatedByID int64          `gorm:"not null"` // References: users(id)
+	CreatedByID int            `gorm:"not null"` // References: users(id)
 	Tags        string         `gorm:"not null"`
 	NoteType    NoteType       `gorm:"not null"`
 	ContentSize int            `gorm:"not null"`
